pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
]
